requirements.txt
Brotli==1.0.9
certifi==2020.12.5
click==7.1.2
cycler==0.10.0
dash==1.20.0
dash-bootstrap-components==0.12.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
mlxtend==0.18.0
mysql-connector-python==8.0.24
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
protobuf==3.15.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
wordcloud==1.8.1
